Mods to build system to get CUDA working
enable the user to include a dice_utils_module_commands.sh file in the
build directory which gives shell commands to load or unload modules
so that nvcc-wrapper can be switched off for the utils build (because
it breaks things like boost::gil, etc.). The user can also include a
dice_module_commands.sh file in the build directory to do custom shell
stuff when the main dice project is built. If neither of these files
exist, these module steps are skipped.