Account for different data layouts in image constructor by array
If an array is passed in, it is assumed to be LayoutRight or row major.
If the default layout in Kokkos is LayoutRight, the array pointer is
used, otherwise if the default layout is LayoutLeft, the data in the
array is copied into the right layout.

For the tiff reader utils, there is a new flag that denotes layout
left/right which manages how the data is populated.