Re-work the external project again to accomodate nvcc needs
The shell scripts approach didn't work very well because cmake
forks the processes that run the configuration. This meant that
environment changes through loading and unload modules may or may
not work. The current implementation requires that the user create
a shell script that builds each part of the project manually,
changing the environment where necessary.