Add offsets to mask so that they are always in global coords
If the user applies a mask to a sub-image or an image that was constructed
as a portion of a larger image, the mask will be assumed to be in the
coordinates of the larger image and the offsets will be applied. That
way mask geometries are defined once at the beginning instead of
for each portion image.