New test for DIC challenge 5 with ZNSSD
Also fixed an issue with parameter lists and scalar_t: Now, all
parameters are of type double and double is the type of the member
variables in the classes that use the parameters. Casting from double
to float or vice versa was messing up parameter access because they
are templated on the return type.