First draft of an initializer class that uses path files
A path file contains potnential solutions for the deformation parameters
and serves as a starting point for the DIC convergence. When a path file
is supplied, it gets filtered to a resolution of 0.5 pixels for displacement
and 0.01 radians for rotation. Then the file gets compressed to a set of
unique points that serve as the set of potential solutions.

This is useful in a few ways: 1 it provides a way to determine if a
subset has become lost. It also provides a way to continue tracking
if one frame fails. Lastly, it greatly improves the initialization of
the DIC convergence routines because the first guess is much more accurate.