Split the core library into core + base libraries
This was to resolve a shared_library dependency on windows. Schema
(in dicecore) depends on dicefft which depends on diceimage (also in
dicecore). Now all the basic DICe classes (Image, Subset, Simplex,
Multifields, etc.) are in library dicebase.

The higher level classes like schema and initializer ... are in library
dicecore still.