Major changes to cmake build system (lib and bin output dirs)
Set all library, runtime, and archive directories to be in a master
lib or bin dir the in the build folder. This puts all the libs and
executables in one place, even if the 'install' target is not made.

Started an external code example that links against DICe libraries
by using an externalproject_add() in cmake. This was part of the
motivation for putting all the libs and execs in one place.