Several minor changes to address issues with threaded branch of code
CMake changes to get custom app working (DICE_MPI flag was not getting set)

Enable the Tpetra/Kokkos variants for the custom app example.

Loosened tolerances for examples with threading, some of the examples
report variables that have values close to zero so the relative diffs
were large. This should be addressed in a more systematic way in the diff
tool at some point.

Added missing call to default_constructor_tasks() in schema constructor
that takes a parameter file name.