Address issue #20 regarding access to image intensity values
Operator () access is const only. The user cannot change the intensity
values using the () accessor.

The intensities() method of DICe::Image is non-const. The user can
change the intensity values using this method (for both the Kokkos
and serial versions).

The best way to change the intensity values is to use the
replace_intensities() method, which also allows the user to re-compute
the gradients. The gradients are not recomputed for the intensities()
method.