Address issue #10 add support for .jpeg and .png
Refactor of image IO to reduce code duplication for tiff, jpeg, ...

The type of file that is written from DICe::Image is now automatically
determined from the file name rather than having to call write_tiff()
or write_jpeg() explicitly. Only one write method now: img->write().

LibTiff is still required, but jpeg and png are optional. If the libraries
for jpeg and png are in the system path, these formats are enabled automatically.
Otherwise the JPEG_DIR or PNG_DIR variables have to be set in the cmake
configure script for them to be optinally included.