Changed the way the user specifies the frame range for a cine file
Now the indexing is done so that it corresponds to the cine's internal
indexing, for example the first frame is now the trigger frame rather
than zero. Even if the trigger frame was -100 before, the user would
still set the cine_start_index to 0 which was confusing. Now the user
would request cine_start_index -100.

Also the output frame numbers in the text files reflect this change