Address issue #33 filter out failed cine pixels at user's request
Sometimes the high speed video can be corrupted because of hardware issues.
This leads to failed pixels in the image frames that have the maximum
intensity value. The user can now filter out these failed pixels and
replace them with the neighbor pixel value. This filter is not the
default and must be requested by the user by setting the
filter_failed_cine_pixels param. A warning is printed that provides
the number of failed pixels for that frame.