Refactor of TEST_FOR_MOTION
Since motion windows are useful for more than just testing if motion
is happening, the keyword has been refactored so that the user
can create a MOTION_WINDOW around a conformal subset to narrow
the amount of a cine frame that needs to be read in (still have to
update the tracking routines that depend on images rather than .cine files).

If the old behavior of TEST_FOR_MOTION is required (check to see if
anything is moving in the frame before analyzing) the motion window
should be created with MOTION_WINDOW <start_x> <start_y> <end_x> <end_y>,
but another keyword needs to be added TEST_FOR_MOTION. Defining the
window is no longer enough to check for motion before analysis.