Refactor of SKIP_SOLVE keyword
Now the user can add either a set of ids after the keyword to specify
frames to turn the subset on or off or a filename that holds the ids.

The first id turns tracking off, subsequent ids turn tracking on and
off in alternating fashion.  For example if the keyword is used like this:
SKIP_SOLVE 12 15 18 27, solves will be executed until frame 12, then
skipped for 12 through 14. Frames 15 through 17 will be solved. Solves will
again be skipped for 18 to 26. Frames 27 and on will be solved.

To get the same behavior, the user could use the following keyword
SKIP_SOLVE subset_0_skips.txt
Where the file subst_0_skips.txt has the syntax
12
15
18
27