GUI: Catch dice executable exceptions
There's no way to get at the exception from the QProcess, but for
the common exceptions in the executable, the executable itself can
report the exception and then return a non-zero exit code. We'll need
to add most of the common exceptions to those reported by the main
executable for this to work. Otherwise the user has no idea what went
wrong.