Change the way BCs are applied
During the first iteration, BCs are applied by setting the residual
to the prescribed value and the tangent row for that dof is zeroed
out (except for a 1.0 on the diagonal). In subsequent iterations, if
the fixed point method is used, the residual is just zeroed out for
each BC dof.