Enable the gradient-based alg to work well for large theta and ref grads
In a recent commit, we enabled the user to use the deformed image gradients
to track large rotations, but the gradients had to be computed for
each frame. In this commit, the gradients are computed only for the
reference frame, but transformed to the deformed orientation. This makes
the compute time much faster and about the same level of robustness.