final touches on parallel mapping for fields to respect seeds and obstructions
Tests now passing in serial and parallel. The new philosophy behind
parallel decomp is that the this_proc_gid_order_ vector is used to
define the order in which the subsets are computed, but does not
dictate the parallel decomp maps. The maps are now ordered in ascending
order regardless of the order in the this_proc_gid_order_ vector.

Had to re-bless some of the parallel tests involving seeds because it
was discovered that the old way of doing things had a bug: the local
id was being stored for the neighbor index in parallel rather than the
global id. The global id is now correctly assigned.