refactor of post processors
The goal of this refactor was to get rid of the dependence on a schema
or assuming that the grid of points is spaced regularly. Now the post
processors just read in a set of points with no knowledge of elements
or connectivity or the need for subset step size, etc.

The field names were also generalized so that one could use an alternate
field than the standard disp or coord fields for the calcs.

Also, got rid of Keys 4 strain post processor