refactor: depricate num_images options in favor of last_image_index
When the automagic input format was used, the num_images option
was not clear. Changed the option to be last_image_index. Now the
automagic iterates files from the reference index to the last_image_index.

Also updated the deformed subset images to index according to the
frame index rather than the frame number. So if the cine file starts
at -100, the deformed subsets image will be <>_-100.tif instead of
<>_000.tif