initialization of the cross-correlation
Using a projective transform to get the initial guess, then using
a simplex optimization to get Q and R for the cross-correlation.
As the algorithm works now, the user will have to give some point
data for matching points in the left and right images. Otherwise the
projective search may not converge. This will be changed to make it
simpler eventually.