major changes to how the cross correlation is initialized
for most problems dealing with DIC on a planar surface, the projection
initialization works fine. For others (for example using satellite
imagery) the projective transform is not enough to get a close enough
for the initial guess of the cross-correlation displacements q and r.
In that case, the user can use the use_nonlinear_projection option
to include an additional warping applied to the image on top of the
projective transform.