mods to get parallel working properly with recent changes to post processors
also fixed some issues with the way the stereo schema is constructed.
No longer doing this manually, but taking as much member data
as possible from the left schema so that all the parallel maps
match for both the left and right. Fields can now be shared between
the left and right because the array maps are identical now.