major schema refactor
to improve the memory use when DICe is executed in parallel, the schema
had to be changed dramatically in how the class is constructed. Previously,
the constructors took pointers to images or image file names. The
schema would then be initialized at a later stage. Now the schema
is constructed with the subset locations or pointers to parameters and
the images are set at a later stage. This way, when the images are loaded
they can be decomposed in parallel. This places a new responsibility on the
app developer to always set the reference and deformed image after the
schema is constructed.

Also reworked how a set of points handed to a schema gets split
across processors in these commits.