major refactor of how cine files are loaded and frame indexing in a schema
now cine files can be loaded like any other image, there is some
name mangling that has to take place where the frame number or
range is added to the file name with the following convention:
to load frame 37 of cine video file cine_file.cine, the string name
"cine_file_37.cine" gets passed to the image constructor that takes
a string argument. Similar conventions can be used to load a set of
frames averaged together. The name would be "cine_file_avg37to51.cine"
In the background a singleton class is created that stores the cine
headers for any video files that get oppened. This way the header is
always available and does not have to get loaded for every frame.

Also had to change how frame indexing is done in a schema to make it
a lot simpler. The whole strategy is based on frame_id now rather than
maintaining an image_it and a frame_id which led to lots of confusion.

The dice main executable is now a lot more easy to follow since the
cine logic could be stripped out and the frame indexing is more natural.