reworked the cmake scripts so that parallel builds work with external projects
historically, the image utils libraries needed to be built as an external
project so that a mix of the nvcc compiler and gcc could be used. The
custom_app example also needed to be built as an external project so
that it was an isolated component. This commit adds the necessary dependencies
in cmake to allow parallel builds (with -j <num>) to function properly