refactor parallel decomposition to avoid interlock for large proc counts
The old strategy for parallel decomp worked up to 40 or so processors,
but was then causing an interlock (and a seg fault). This was due to too much data being passed
all-to-all regarding the subset coordinates. As it has been redone,
proc 0 now partitions the mesh and communicates selected portions to
each processor. No more all-to-all, or excessive copies of the coords
vector. The parallel decomp should respect obstructions for the
tracking routine and neighbors when USE_NEIGHBOR_VALUES is active for
the initialization routine.