consolidated the interface for schema fields and mesh fields
now the schema interface to mesh fields is dictated by the same
field specs that define the underlying mesh fields. The old way
of having enums for the schema and field specs for the mesh fields
led to a lot of confusion. The old enums where removed and replaced
with DOF_<> enums to make sure that a code developer used to the old
way of doing things would get a compile error if they tried to
use the old style of identifying fields. The interface is a lot clearner
now and easier to understand.