enable 2D global DIC in the GUI
Mostly go the live_plot post processor working for global DIC so
the user can see the line plots of the results data while the analysis
is in progress. Also needed to relax some of the constraints in the BC
manager so that it did not require BCs to be defined in the subset
file, it now assumes all the boundaries are stress free if there
are no BCs defined.