Enable user to specify a user-defined mesh for global method
Now if the user has a custom 3-node triangle mesh they want to use
for the global method, they can specify in the input file a new
keyword: mesh_file with the path to the mesh. The format for the mesh
is simply all the nodes listed by coordinates with a 'v' at the start of
the line to denote a vertex, for example:

V 1.25 3.45 0.0
(note the three coordinates with zero for the Z-dimension)

then the connectivity for each element listed with an 'f'

F 23 54 89

Also added a test to the regression suite global_user_mesh that has
an example of the input syntax, as well as an example mesh file.