cross projection and 3D rotation/translation functions added to CamSystem as well as pass throughs for the camera parameters
The functions for the projection shape functions have been added to CamSystem.
Function cross_projection_map uses the individual projection functions in the camera class to project subset locations from one camera to another it is overloaded for single points, vectors of points and first partials.
Function rot_trans_3D takes 3D world coordinates and applies a rotation/translation transformation to them. Also overloaded for single points, vectors and first partials.
Also added to the class are functions to add and remove cameras to the system and to fill the governing parameters of the cameras. Now cameras can be added and populated without having to save the data in a file.