updates to matrix class to enable runtime sizing
the matrix class is intended to be sized at compile time for performance
reasons. Also, the storage for the matrix class is on the stack so
it's limited to small dimensions (currently the max is 16x16). In
some cases, the size of the matrix is not known until runtime
so a runtime sizing method was added to the matrix class. When sized
at runtime, the storage is MAX_MATRIX_SIZE x MAX_MATRIX_SIZE, but the
active dimensions are the runtime dimensions.