refactoring of Camera and Camera_System classes
note: to keep this commit more concise I didn't update the stereocalib2 files
to include the changes to the Camera and CameraSystem classes, I temporarily
removed them from the CMake lists so they won't get compiled. I also
removed the TestStereoCalib2 test, which depends on these classes. I will
re-enable them in the next commit in which I will update the Stereo_Calib class.

Consolidated many of the Camera and Camera_System methods so that code is
not replicated between methods. Also, made the interface to their constructors
more clean by introducing a Camera_Info struct.

Added more protection to the Camera classes member data, I took the approach
that once a camera is constructed, it's parameters won't change so I
removed a lot of the set methods. (It's helpful to have these methods
in testing, but they won't be used much in practice so I removed them.)
There is still a mechanism for changing the camera parameters and that is
to call the constructor again with a new Camera_Info argument.

Removed the idea of having a fixed source and target camera and changed
these to arguments in the projection methods. That way when we go beyond
two cameras, this class will not need to be updated.

Added a lot more testing on array bounds.