a new calibration class to merge the privious stereo calib classes/free functions
this new class also provides for single camera calibration
consolidated a lot of the input/output functions for the old
stereo calib classes

updated the tests to be more meaningful and to check if running the cal
with finding the intersection points produces the exact same result
as running the cal with intersection points read from a file

simplified the interface to running a calibration, the user only
needs to put parameters in a file or parameterlist and run the
calibrate() method. The return value is either the rms error or a camera
system object.

works with checkerboard or dot targets

changed the debugging output images to use color to make it easier to
see what's going on with the cal