integration of CameraSystem class into existing triangulation routine
added a utility to add the image width and height to old .txt format
calibration input files. The new DICe xml format will be used going
forward, but this will enable users that have cal files from previous
analyses to be able to use those files.

Rather than have the triangulation class read in the calibration
parameters, it uses a camera system now to do so. Also the storage
of the matrices for a triangulation class are now the new Matrix class.

Put in routines to convert the extrinsic parameters depending on
weather the user specifed them in terms of world to cam or in
reference to camera 0 to camera 1

added more testing in the camera inverse lens distortion initialization
routine that will throw an exception on nan or inf in the calculation

TODO: We need to figure out what lens distortion model VIC3D uses bacause the
k3 value in their xml files is huge and this screws up all of our current
lens distortion models.