generalize the OpenCVServer with utilities that can be used by other classes
this enables the GUI to show preview images of what the calibration
routines are doing, both of which using the same utilities in the
OpenCVServerUtils files

Also, created a new test of the OpenCVServer

refactored the server to turn the input string into a parameterlist and
pass that parameter list to the opencv methods. This makes the
requested filters a lot more clean.