mods to facet projection as it relates angles theta, phi, and length Zp to the camera params
for calibrated results in 2d, the camera's rotation matrix and translation vector
describe a transformation from the coordinate system that lines up with the surface
of the calibration plate in physical space with the camera's coordinate system. Now
this information can be used to determine the angles theta and phi and length Zp
that define the plane corresponding to a facet projection.

also added a component test for the porjection from image coordinates to world
and back with synthetic images DICe_TestImageToWorld