do neighbor search for VSG based on subset coordinates not model coordinates
*** This commit will cause changes to strain values from VSG in stereo,
it's a major bug fix.

The parameter that gets set in the params.xml file is strain_window_size
in_pixels. The way that parameter was used prior to this commit in
stereo analysis was to as the physical size of the window in model
coordinates. This was confusing. Now the window size in pixels is
enforced based on the subset coordinates in pixels of the left image
when finding neighbors for the VSG calcs.

Also fixed a bug: the neighborhoods were being established before
the right fields were set. This is now fixed