fix test issue for exodus files that fail on creation
depending on which version of trilinos you build against, if you
try to create an exodus file when one with the same name exists
it throws an error. The GlobalPixelShift test was creating an exodus
output file with the same name for several cases represented in the
test. This was causing test fails on linux so I commented out the
exodus file creation (it's not actually used in the test).