move image IO utilities out of external project and into main cmake
The image io utilities were originally carved off into an external
project to facilitate building them with CUDA and to enable using Kokkos.
Since Kokkos has been stripped out, it's no longer necessary to build
the io utilities as an external project.

It was also a pain to have to 'make clean' whenever you changed
something in the io utilities. If you forget to do this your changes
wouldn't be included. Now, since the io utilities are part of the
main project they get built whenever the code is changed.