remove old cine file reader (cine.cpp/cine.h) in favor of hypercine lib
hypercine allows for reading large chunks of a cine file into a memory
buffer. The dimensions of these chunks are defined in an object called
a hyperframe which is a collection of frames (that don't have to be in
order or continguous) and windows inside the image boundary. Reading
these blocks avoids having to open and close the cine file for each
file read. There was a lot of work that had to be done to the DICe::
Image class because previously it would re-allocate a new block of
memory for each image regardless of the type. Now for cine images
it only allocates the image memory if the hypercine hasn't already
done so in its memory buffer. The goal is to make reading of cine
files extremely fast, particularly in application where only a small
portion of the image frame is actually used.

One other major change that this commit made was the base type for
image intensity values. In the past, they were float values, but to
save memory and to enable dealing with image memory by passing a pointer
around instead of copying the image data, the base type for intensity
values is now uint16_t. This enables images to be stored in less memory
and improves cache performance.