first commit of Image class refactor to be templated on storage type
in some cases the image intensity storage type is best suited as
an a low memory int type like uint16_t. In other cases the user might
want more precision, for example with a synthetic image constructed
using a functional representation. In those cases the truncation
become problematic. To enable both of these cases the Image class
is now templated on a storage type which defaults to storage_t
which is set in DICe.h.