enable either a deep copy of cine intensity values or pass pointer
an image param can be set (buffer_persistence_guaranteed) that lets
the image constructor know that a pointer to the cine reader buffer
can be used for the intensity values and that updating that buffer
later will not cause issues with the original image. To do this, all
motion windows or extents of the image have to be known and read into
the buffer ahead of time (as is the case with the tracking routines).

In other cases, where one sub region of a frame may be read before another,
the reading of the second sub region will overwrite the first one in the
buffer. An error gets thrown if the persistence flag is used and the
buffer gets overwritten.