resolve multiple hypercine singleton issue on windows build
turns out singletons are local to .dll files on windows so if you
call an instance of a singleton from one .dll in another you get
another instance of the "singleton". To prevent this from happening
several helper functions were added to the diceioutils library so
that calls from other libraries don't need to get an instance of
the hypercine singleton directly.