add an additional check for blob size=500 to calibration keypoint finding
In some cases, the min blob size was too big (100) so a secondary check
was added for blobs of size 10 several months ago. Now it seems that adding
a third pass with a larger blob size is a good ideas since large blob sizes
are possible for larger or high resolution images. A third pass with
min_blob_size=500 has been added.