added new initialization method for cross-correlation: USE_SPACE_FILLING_ITERATIONS
Rather than try doing an initial projection of the image from a
plane in the left frame of reference to the right, which works well
for planar objects, take another approach. In the space filling approach
an initial feature matching is done, then these seed points are
cross-correlated. The rest of the points are correlated in a
cascading fashion starting with the seeds. There are checks on the
correlation sigma values and epipolar error for each point to ensure
good seeds are used. This correlation initialization method works
better for non-planar surfaces.