reset lost subsets for FEATURE_MATCHING initializer
Previously, if a subset got lost during a frame leading to SIGMA = -1.0
the subset would be skipped for subsequent frames. Now if the FEATURE_MATCHING
initializer is used, the subset is reset after each frame in case a lost
subset re-enters the field of view, it will now be analyzed.