change how SIGMA is calculated for partial subsets
When a subset partially extends out side of the image boundary, the
way SIGMA was calculated previously would decrease SIGMA for these
subsets because of the fewer pixels, but intuitively SIGMA should increase
for these subsets. Now the noise standard deviation calculations
account for the deactivated pixels which fixes this issue.