minor test changes for linux (remove unicode test and fix tols)
The unicode filenames don't get translated correctly on all linux
systems, as a result the filenames come up as question marks in the
regression test images. This commit turns the unicode filename
test off on linux.

For tests like the reset_lost_subsets test, the solution for the
last step is in the noise (10E-9) setting a floor value for this
test so the comparison for small values doesn't fail the test.