corrected for lens distortion with the epipolar constraints
The epipolar constraint is used as a tolereance to check for valid
points when doing the cross correlation. Now that constraint takes
lens distortion into account, previously lens distortions would
lead to large epipolar errors and the cross correlation initialization
would fail.

Also, added a new test in the regression tests for this.