take another approach to random access frames in video reader
Depending on which version of OpenCV is linked into DICe one can
see very different behavior when trying to access video frames
from an mp4 file out of sequence. To fix this DICe now accesses
frames by time stamp, not frame id (unless a sequence of images
are being processed, then DICe simply reads the frames in
succession). Part of the issue that since mp4 is a compressed format,
accessing a particular frame requrires reading the nearest keyframe
and applying deltas to get the frame of interest. Accessing the
frame by time stamp seems to fix this issue.