tool for rectifying images
testing out new methods to initialize the cross-correlation. Many of
the methods require stereo rectified images so a new tool has been
created that takes to input images and a calibration file and
outputs a pair of stereo rectified images.

Also added some methods to Triangulation to access things like the
camera matrix or the distortion coeffs.